requirements.txt
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
mccabe==0.6.1
newsapi-python==0.2.6
numpy==1.19.1
pandas==1.1.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.1.0
zipp==3.1.0
